requirements.txt
# Core dependencies
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
requests>=2.31.0

# Data processing
rapidfuzz>=3.5.0
openpyxl>=3.1.0
unidecode>=1.3.0
plotly>=5.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
python-dotenv>=1.0.0

# Optional AI/NLP (if needed later)
# spacy>=3.7.0
